feat: Expose DataPicker without the FloatingContainer
