+++ src/DateRangePicker/Calendar.tsx
@@ -18,7 +18,7 @@ import startOfDay from 'date-fns/start_of_day'
 import startOfMonth from 'date-fns/start_of_month'
 import * as React from 'react'
 import { IconAngleLeft, IconAngleRight, IconButton } from './Calendar.styles'
-import { AnyDate, DateRange } from './DateRangePicker'
+import { AnyDate, DateRange } from './DateRangePicker.types'
 import { checkEndEdge, checkRange, checkStartEdge, isOutsideMinMax } from './dateUtils'
 import DayCell from './DayCell'
 import { defaultClasses, months } from './enums'
